+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.3.2-pre
 * Add build reports to readme
 * Fix broken build command when no parameter is specified
+* Fix broken auth username and password prompt with interactive flag
 
 ### 0.3.1
 * Monitor Jenkins instance and view checks all jobs instead of just the first job